Update pom.xml
The version number was erroneously changed to 2.5.x but the current version is actually 2.4.1 (as the software will tell). Changes related to integration in Fiji (such as pom.xml) should not actually change the version number of the imagescience library itself, or at most the tertiary version number, so it was changed back.